backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pdf2image==1.16.3
pillow==10.4.0
python-multipart>=0.0.7
opencv-python>=4.8.0
numpy>=1.24.0
torch==2.9.1
torchvision==0.24.1
